Move patch_retry_predicate() to its own module
As the function is not specific to PublisherClient and can be reused
in e.g. SubscriberClient, it is extracted out of the publisher_client
module to a common import place.
